+++ R/events_tbl.R
@@ -51,14 +51,14 @@ new_events_tbl <- function(.id = integer(0),
   }
   if (!is.null(sampling_rate)) {
     events[, .initial := sample_int(as.integer(.initial),
-      sampling_rate = sampling_rate
-    )]
+      sampling_rate = sampling_rate)
+      ]
     events[, .final := sample_int(as.integer(.final),
-      sampling_rate = sampling_rate
-    )]
+      sampling_rate = sampling_rate)
+      ]
   }
   data.table::setattr(events, "class", c("events_tbl", class(events)))
-  events
+  events[]
 }
 as_events_tbl <- function(.data, ...) {
   UseMethod("as_events_tbl")